Split into packages and add readme
