+++ pkg/acceptance/bettertestspoc/assert/commons.go
@@ -1,4 +1,4 @@
-package bettertestspoc
+package assert
 
 import (
 	"errors"
