+++ pkg/acceptance/bettertestspoc/assert/resource_assertions.go
@@ -1,4 +1,4 @@
-package bettertestspoc
+package assert
 
 import (
 	"errors"
