+++ pkg/acceptance/bettertestspoc/assert/snowflake_assertions.go
@@ -1,4 +1,4 @@
-package bettertestspoc
+package assert
 
 import (
 	"errors"
