+++ pkg/acceptance/bettertestspoc/assert/warehouse_parameters.go
@@ -1,4 +1,4 @@
-package bettertestspoc
+package assert
 
 import (
 	"strconv"
