+++ pkg/acceptance/bettertestspoc/assert/warehouse_parameters_ext.go
@@ -1,4 +1,4 @@
-package bettertestspoc
+package assert
 
 func (w *WarehouseParametersAssert) HasDefaultMaxConcurrencyLevel() *WarehouseParametersAssert {
 	return w.
