+++ pkg/acceptance/bettertestspoc/assert/warehouse_resource.go
@@ -1,4 +1,4 @@
-package bettertestspoc
+package assert
 
 import (
 	"testing"
