+++ pkg/acceptance/bettertestspoc/assert/warehouse_show_output.go
@@ -1,4 +1,4 @@
-package bettertestspoc
+package assert
 
 import (
 	"strconv"
