+++ pkg/acceptance/bettertestspoc/assert/warehouse_snowflake_ext.go
@@ -1,4 +1,4 @@
-package bettertestspoc
+package assert
 
 import (
 	"fmt"
