+++ pkg/acceptance/bettertestspoc/config/warehouse_model_ext.go
@@ -1,4 +1,4 @@
-package bettertestspoc
+package config
 
 func BasicWarehouseModel(
 	name string,
